+++ frontend/utils/lazyLoadedView.js
@@ -1,4 +1,4 @@
-import { Vue } from '@common/common.js'
+import Vue from 'vue'
 import ErrorModal from '@views/containers/loading-error/ErrorModal.vue'
 import ErrorPage from '@views/containers/loading-error/ErrorPage.vue'
 import LoadingModal from '@views/containers/loading-error/LoadingModal.vue'
