+++ frontend/views/containers/contributions/PaymentMethods.vue
@@ -50,9 +50,10 @@ fieldset(data-test='paymentMethods')
 <script>
 import { mapGetters } from 'vuex'
 import { validationMixin } from 'vuelidate'
-import { Vue, L } from '@common/common.js'
+import { L } from '@common/common.js'
 import { maxLength, required } from 'vuelidate/lib/validators'
 import { GROUP_PAYMENT_METHOD_MAX_CHAR } from '@model/contracts/shared/constants.js'
+import Vue from 'vue'
 
 export default ({
   name: 'PaymentMethods',
