+++ frontend/views/containers/payments/RecordPayment.vue
@@ -70,7 +70,7 @@ modal-base-template(ref='modal' :fullscreen='true' class='has-background' v-if='
 
 <script>
 import sbp from '@sbp/sbp'
-import { Vue, L } from '@common/common.js'
+import { L } from '@common/common.js'
 import { mapState, mapGetters } from 'vuex'
 import { PAYMENT_PENDING, PAYMENT_COMPLETED, PAYMENT_NOT_RECEIVED, PAYMENT_TYPE_MANUAL } from '@model/contracts/shared/payments/index.js'
 import { validationMixin } from 'vuelidate'
@@ -78,6 +78,7 @@ import SvgSuccess from '@svgs/success.svg'
 import { dateToMonthstamp } from '@model/contracts/shared/time.js'
 import ModalBaseTemplate from '@components/modal/ModalBaseTemplate.vue'
 import RecordPaymentsList from '@containers/payments/RecordPaymentsList.vue'
+import Vue from 'vue'
 import BannerScoped from '@components/banners/BannerScoped.vue'
 import BannerSimple from '@components/banners/BannerSimple.vue'
 import ButtonSubmit from '@components/ButtonSubmit.vue'
