+++ frontend/views/containers/proposals/AddMembers.vue
@@ -46,7 +46,7 @@ proposal-template(
 
 <script>
 import sbp from '@sbp/sbp'
-import { Vue } from '@common/common.js'
+import Vue from 'vue'
 import { mapState, mapGetters } from 'vuex'
 import { validationMixin } from 'vuelidate'
 import { PROPOSAL_INVITE_MEMBER } from '@model/contracts/shared/constants.js'
