+++ frontend/views/utils/allowedUrls.js
@@ -1,4 +1,4 @@
-import { Vue } from '@common/common.js'
+import Vue from 'vue'
 
 /**
  * Table of all well-known safe URLs used in the app UI.
