+++ frontend/views/utils/vError.js
@@ -1,4 +1,4 @@
-import { Vue } from '@common/common.js'
+import Vue from 'vue'
 
 // Register a global custom directive called `v-error`
 // to automatically display vuelidate error messages
