+++ frontend/views/utils/vStyle.js
@@ -1,4 +1,4 @@
-import { Vue } from '@common/common.js'
+import Vue from 'vue'
 /*
 * Style tag overload because Vue is trying to compile
 * the content inside the tag otherwise
